refactoring pod api
