--- dynamiczne/progress/WWL21221.csv
@@ -1,4 +1,0 @@
-Wyjazd,1,12.12.12,0.0,1090.0
-Tankowanie,1,13.12.12,100.0
-Powrót,1,22.12.12,2900.0,390.0
-Norma,1,22.12.12,27.586206896551722
