--- dynamiczne/progress/WWY12112.csv
@@ -1,5 +1,0 @@
-Wyjazd,1,12.12.12,0.0,1090.0
-Tankowanie,1,13.12.12,190.0
-Tankowanie,1,19.12.12,290.0
-Powrót,1,20.12.12,3690.0,390.0
-Norma,1,20.12.12,31.978319783197833
