+++ dynamiczne/progress/WWY21221.csv
@@ -18,3 +18,7 @@ Wyjazd,5,06.02.12,11000.0,1090.0
 Tankowanie,5,07.02.12,160.0
 Powrót,5,13.02.12,13000.0,680.0
 Norma,5,13.02.12,28.499999999999996
+Wyjazd,6,14.02.12,13000.0,1090.0
+Tankowanie,6,15.02.12,190.0
+Powrót,6,22.02.12,16900.0,195.0
+Norma,6,22.02.12,27.82051282051282
